requirements.txt
json
argparse
requests
pycryptodome
